Task 13 jpa db (#25)

+++ app/src/main/java/pl/futurecollars/invoicing/controller/invoice/InvoiceApi.java
@@ -23,17 +23,17 @@ public interface InvoiceApi {
 
     @ApiOperation(value = "Add new invoice to system")
     @PostMapping
-    ResponseEntity<Integer> saveInvoice(@RequestBody Invoice invoice);
+    ResponseEntity<Long> saveInvoice(@RequestBody Invoice invoice);
 
     @ApiOperation(value = "Get invoice by id")
     @GetMapping("/{id}")
-    ResponseEntity<Invoice> getInvoiceById(@PathVariable int id);
+    ResponseEntity<Invoice> getInvoiceById(@PathVariable Long id);
 
     @ApiOperation(value = "Update invoice with given id")
     @PutMapping("/{id}")
-    ResponseEntity<?> updateById(@PathVariable int id, @RequestBody Invoice invoice);
+    ResponseEntity<?> updateById(@PathVariable Long id, @RequestBody Invoice invoice);
 
     @ApiOperation(value = "Delete invoice with given id")
     @DeleteMapping("/{id}")
-    ResponseEntity<?> deleteById(@PathVariable int id);
+    ResponseEntity<?> deleteById(@PathVariable Long id);
 }
